pyproject.toml
[tool.poetry]
name = "ohme-client"
version = "0.1.0"
description = ""
authors = ["mnbf9rca <mnbf9rca@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.9,<3.12"
httpx = "^0.25.1"
python-dotenv-vault = "^0.6.3"
jsonschema = "^4.19.2"


[tool.poetry.group.test.dependencies]
pytest-black = "^0.3.12"
flake8 = "^6.1.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
coverage = "^7.3.2"
pytest-mock = "^3.12.0"
mock = "^5.1.0"
flake8-pyproject = "^1.2.3"


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = [
    "*/test/*",
    "*/tests/*",
    "*/venv/*",
    "*/.venv/*",
    "*/.tox/*",
    "*/.eggs/*",
    "*/.vscode-remote/*"
]
relative_files = true

[tool.pytest.ini_options]
addopts = "--cov --cov-branch --cov-report term-missing -vv --strict-markers -rfE"

[tool.flake8]
max-line-length = 120
exclude = [
    ".git",
    "__pycache__",
    "docs/source/conf.py",
    "old",
    "build",
    "dist",
    "venv"
]
